backend/requirements.txt
flask==3.0.0
authlib==1.3.0
werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
pyjwt==2.8.0
cryptography>=41.0.0,<46.0.0
boto3>=1.34.0
botocore>=1.34.0
